requirements.txt
pytorch_lightning
torchmetrics
h5py
scipy
scikit-learn
matplotlib
seaborn
easydict
wandb